requirements.txt
--extra-index-url https://builds.golem.network
setuptools>=36.0.1
appdirs>=1.4
click>=4.0
docker-py==1.7.2
netifaces==0.10.4
peewee<2.11,>=2.8
peewee_migrate==0.14.0
Pillow==4.2.1
psutil==5.2.2
pluggy==0.6.0
asn1crypto==0.22.0
pyasn1==0.4.1
pycryptodome==3.4.11
pysha3>=1.0.2
pytz
six==1.10.0
Twisted==17.9.0
cffi==1.10.0
ethereum==1.6.1
eth-abi==0.5.0
eth-keyfile==0.4.1
eth-keys==0.1.0b4
eth-utils==0.7.4
eth-tester==0.1.0b15
web3==3.16.4
autobahn==17.10.1
cbor2==3.0.4
base58
ndg-httpsclient
certifi
tabulate
PyYAML>=3.12
crossbar==17.12.1
txaio==2.8.2
pydispatcher
Crypto
scrypt
packaging
semantic_version
pyOpenSSL==17.1.0
requests==2.18.4
Automat==0.6.0
zope.interface==4.4.2
incremental==17.5.0
OpenEXR
pycparser==2.17
numpy==1.13.1
scipy==0.19.1
pyssim
enforce==0.3.4
miniupnpc<2.1,>=2.0
token_bucket==0.2.0
py-cpuinfo==3.3.0
humanize==0.5.1
raven
Golem-Messages==1.17.2
-e git://github.com/golemfactory/golem-smart-contracts-interface.git@07c939221967cd9edb42dddba7cef460b905d953#egg=golem_sci
html2text==2018.1.9
